tests/mock-packages/kitchen-sink/pyproject.toml
[project]
name = "99...S1Lly---namE___PACKAG3-.x-_y_.z"
description = "🚀 🦄 \"quoted\" 'quoted again' $$$"
authors = [
    { name="Brénainn Woodsend", email="bwoodsend@gmail.com" },
]
dynamic = ["version"]
license = { file="The license file" }
dependencies = ["certifi>=2020.4", "pywin32-ctypes; platform_system == 'Windows'", "tzlocal"]

[build-system]
requires = ["colorama; platform_system == 'Linux' and python_version >= '3.8' and implementation_name == 'cpython'", "setuptools", "setuptools-scm"]

[project.urls]
"Homepage" = "https://github.com/bwoodsend/polycotylus"

[tool.setuptools]
py-modules = ["sIlLy_Nam3"]

[tool.setuptools_scm]
